generator/gen.go: syscall group description
Since systemd's new output format, a one-liner describing each syscall
group is available as a comment after the group's name. The parser was
adjusted to catch this description and store it with the set.

Following, this comment is put next to each map entry in the
syscallset_linux.go file. Furthermore, a new generator/go-apidoc.tmpl
was introduced to create a package documentation for this package
listing all known systemd syscall groups.